+++ Utils/UiModeStore.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Text;
 using System.Text.Json;
 
 namespace IspAudit.Utils
@@ -62,7 +63,7 @@ namespace IspAudit.Utils
                 };
 
                 var json = JsonSerializer.Serialize(state, new JsonSerializerOptions { WriteIndented = true });
-                File.WriteAllText(path, json);
+                FileAtomicWriter.WriteAllText(path, json, Encoding.UTF8);
             }
             catch
             {